Change ADT structures and sketch event module
Now, instead of having just linked lists for everything, they are only
kept for the scree list, as they are usually a fixed small number.
Desktop list is now a doubly linked circular list, kept as a list
because it's not a big deal.  Desktops are typically relatively small in
number and they are set only on the beginning of the configuration.
Windows, on the other hand, is chaotic, and a list is not going to
perform well, so it's changed to an open-addressed hash table (closed
hashing).  I hope this changes will improve performance.

Some other changes were made, specially on the event loop.  The update
functions are now implemented and the hash table will reduce the
complexity a lot when traversing all the screens, desktops and windows.
As for events, the main structure skeleton is almost done.