Change window manager structure to singleton
The window manager structure (`wm_td`) is now a singleton structure and
handles the configuration on its own instead of receiving it as
a parameter.  A parameter for display was included and added as
option.  Some other changes were made, and errors were corrected.

Several functions, all of the string-related operations, were rewritten
and added in a new file.  This functions are "safe" in terms that they
check for null pointers before using them, or ending all strings with
the null byte.  In this regard, don't need to use those of `string.h`
and hence, don't need to use the `_POSIX_C_SOURCE` of 2008 for `strdup`,
just `_POSIX_C_SOURCE` of 1990 is fine, just for `getopts`.