Update comments for uniformity
Comments were updated and redacted in a more uniform way to make Doxygen
output more equal.  Even though some of the special commands are aliases
for one another, it's easier to make `grep` searches on them in this
way.  Now, normal emphasis is `@e`, functions and macros are `@a`,
parameters are `@p`, and types or structures are `@c`, just as regular
code.