Refactor and restructured code; from C99 to C11
The code has been refactored to eliminate circular dependencies and to
include forward declarations.  The enumeration for priority levels has
been moved to a separate file, as several functions across different
modules rely on it.  Because of this, the C standard being used now is
C11 and not C99.

Regarding events, we have started to establish a generic
action-event-queue system.  For actions that require additional data,
a pointer to a corresponding data structure is populated with the
necessary information.  Otherwise, the action is simply added to the
priority queue for later processing.