Improve hash table operations
Now, the open-addressed hash table to manage windows is capable or
resizing itself on insertions or removal of items.  If more size is
neeeded because its size is reaching the upper threshold, it doubles in
size, or, if that much space is not needed, it halves the size.  These
operations are O(n) complex, but only invoked specifically when needed.