Refactor of EWMH hints module and commands
There were changes concerning the hints information.  In order to
simplify the `eventq` processing, a few commands will be implemented.
Semantics are very important here, especially when talking about
possible future additions of other hints standards.  The structure must
be the same.  Main goal is to generate complex commands in a simple way
that, when invoked, the update without entanglement X routines, EWMH
routines, etc.

The "commands" will be in a another subdirectory to be divided by object
(screen, desktop, window), and managed from the `cmd` module, that
integrates everything in the `cmds` directory (X11 code to handle
windows and properties) and the `hints` directory (where all functions
to handle hints will be stored).