Replace X11 with XCB for improved asynchrony
Replaced the usage of the X11 library with XCB due to the increased
asynchronicity it offers.  The code has been adapted and refactored
accordingly while maintaining the same function signatures.  Moreover,
as `xcb_icccm` and `xcb_ewmh` are now used, the modules handling the
hints system have been removed, for they are no longer necessary.

Some bugs in the operations of the doubly linked circular list ADT have
been fixed, correcting instances where elements were inserted in the
wrong order, both in `cdlist_ins_prev` and `cdlist_ins_next`.  This
solved a problem where desktops were introduced in the incorrect order
on the surface structure.

Name of base structures were changed, not only 'screen' to 'surface',
but 'window' to 'client' in order to make the code more clear and avoid
semantic conflicts with `xcb_screen_t` and `xcb_window_t` (or,
previously in Xlib, `Screen` and `Window`).  All enumerations of
properties and actions have been renamed accordingly.

Regarding the event priority queue, it now processes its events using
threads that continuously monitor the queue for events to handle,
independently of the main application flow.  Since the purpose of this
event queue is to pass events to XCB, they will be processed faster in
the main loop (at least that's my intention).