Adapt to manage conflicting hint and state logics
Available actions for various objects--such as clients, desktops, and
surfaces--have been established as placeholders to encapsulate the
associated logic.  Some modifications have been made to the client
actions to incorporate the necessary code for EWMH and ICCCM hints;
however, there remains a substantial amount of work ahead.

The implementation of the hint system will delay this project, as it
does not adhere to logical patterns within the window manager.  Instead,
the window manager must adapt to the intrinsic logic of its own nature
whilst simultaneously conforming to external specifications that do not
account for the contradictory states that need management.  Balancing
two distinct logic systems and navigating the interactions between
internal states and those dictated by the hint system is a cumbersome
task.